foodie/requirements.txt
alembic==1.13.2
appdirs==1.4.4
aspose-pdf==24.6.0
async-timeout==4.0.3
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
cryptography==43.0.0
fattura-elettronica-reader==3.0.5
filetype==1.2.0
Flask
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
flask-redis==0.4.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-SSE==1.0.0
fpyutils==3.0.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
idna==3.7
IronPdf==2024.7.1.8
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==4.9.4
Mako==1.3.5
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.1
pdfkit==1.0.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pythonnet==3.0.3
pytz==2024.1
PyYAML==6.0.2
redis==5.0.7
requests==2.32.3
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
eventlet
stripe==6.6.0
Flask-Mail==0.9.1
APScheduler==3.10.4
Flask-Caching


